In shopitems, now the qtyInStock is included
Modified db price's data type to float

+++ backend/model/database.js
@@ -110,7 +110,7 @@ const Product = Singleton.createSingleton.getInstance().define('product', {
         allowNull: false
     },
     price: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.FLOAT(7,2),
         allowNull: false
     },
     prod_description: {
